pyproject.toml
[project]
name = "temporal-workflow-history-analysis"
version = "0.1.0"
description = "Uses AI to analyze workflow histories"
authors = [
    {name = "Rick Ross",email = "rick.ross@temporal.io"}
]
license = {text = "Apache2"}
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "flask (>=3.1.0,<4.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "temporalio (>=1.10.0,<2.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
